+++ CHANGELOG.md
@@ -19,6 +19,14 @@ A component for selecting hour and minute.
 
 They return undefined if parse fails.
 
+### Updated `Select`
+
+- Select component now allows menu component to be wider than the input.
+
+### Breaking changes
+
+- Select options now has `whiteSpace` set to `nowrap`.
+
 ## 7.4.0
 
 ### grid-export